requirements.txt
pyautogui
pynput
pillow
opencv-python
easyocr
numpy